requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiohttp>=3.9.0
feedparser>=6.0.10
deep-translator>=1.11.0
beautifulsoup4>=4.9.1
pytz>=2025.2
Flask>=3.0.0
requests>=2.31.0
Werkzeug>=3.0.0
